chore:兼容firefox

+++ src/giantapp-wallpaper-ui/src/lib/client/api.ts
@@ -10,7 +10,7 @@ class API {
   private emitter = new EventEmitter();
 
   constructor() {
-    if (typeof window === 'undefined' || !window.chrome.webview) return;
+    if (typeof window === 'undefined' || !window.chrome || !window.chrome.webview) return;
     const { api } = window.chrome.webview.hostObjects;
 
     // //注册InitProgressEvent事件
@@ -38,7 +38,7 @@ class API {
     key: Key,
   ): Promise<ApiResult<T>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
       var json = await api.GetConfig(key);
       let res = JSON.parse(json);
@@ -54,7 +54,7 @@ class API {
     value: any,
   ): Promise<ApiResult<null>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
       await api.SetConfig(key, JSON.stringify(value));
       return { error: null, data: null };
@@ -66,7 +66,7 @@ class API {
 
   async getWallpapers(): Promise<ApiResult<Wallpaper[]>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
       var json = await api.GetWallpapers();
       let res = JSON.parse(json);
@@ -82,7 +82,7 @@ class API {
   */
   async getScreens(): Promise<ApiResult<Screen[]>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
       var json = await api.GetScreens();
       let res = JSON.parse(json);
@@ -95,7 +95,7 @@ class API {
 
   async showWallpaper(wallpaper: Wallpaper): Promise<ApiResult<null>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
       await api.ShowWallpaper(JSON.stringify(wallpaper));
       return { error: null, data: null };
@@ -105,25 +105,9 @@ class API {
     }
   }
 
-  // /**
-  //  * @deprecated
-  //  */
-  // async getPlayingWallpaper(): Promise<ApiResult<Wallpaper[]>> {
-  //   try {
-  //     if (!window.chrome.webview) return { error: "no webview", data: null };
-  //     const { api } = window.chrome.webview.hostObjects;
-  //     var json = await api.GetPlayingWallpaper();
-  //     let res = JSON.parse(json);
-  //     return { error: null, data: res };
-  //   } catch (e) {
-  //     console.log(e);
-  //     return { error: e, data: null };
-  //   }
-  // }
-
   async getPlayingStatus(): Promise<ApiResult<PlayingStatus>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
       var json = await api.GetPlayingStatus();
       let res = JSON.parse(json);
@@ -136,7 +120,7 @@ class API {
 
   async pauseWallpaper(screenIndex?: number): Promise<ApiResult<null>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
       await api.PauseWallpaper(screenIndex);
       return { error: null, data: null };
@@ -148,7 +132,7 @@ class API {
 
   async resumeWallpaper(screenIndex?: number): Promise<ApiResult<null>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
       await api.ResumeWallpaper(screenIndex);
       return { error: null, data: null };
@@ -160,7 +144,7 @@ class API {
 
   async stopWallpaper(screenIndex?: number): Promise<ApiResult<null>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
 
       const { api } = window.chrome.webview.hostObjects;
       await api.StopWallpaper(screenIndex);
@@ -173,7 +157,7 @@ class API {
 
   async setVolume(volume: number, screenIndex?: number): Promise<ApiResult<null>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
       await api.SetVolume(volume.toString(), screenIndex + "");
       return { error: null, data: null };
@@ -185,7 +169,7 @@ class API {
 
   async getVersion(): Promise<ApiResult<string>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
       var res = await api.GetVersion();
       return { error: null, data: res };
@@ -197,7 +181,7 @@ class API {
 
   async openUrl(url: string): Promise<ApiResult<null>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
       api.OpenUrl(url);
       return { error: null, data: null };
@@ -209,7 +193,7 @@ class API {
 
   async uploadToTmp(fileName: string, content: string | ArrayBuffer): Promise<ApiResult<string>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
       var data = await api.UploadToTmp(fileName, content.toString());
       return { error: null, data };
@@ -219,25 +203,9 @@ class API {
     }
   }
 
-  // /*
-  // *deprecated
-  // */
-  // async createWallpaper(title: string, coverUrl: string, pathUrl: string): Promise<ApiResult<boolean>> {
-  //   try {
-  //     if (!window.chrome.webview) return { error: "no webview", data: null };
-  //     const { api } = window.chrome.webview.hostObjects;
-
-  //     var res = await api.CreateWallpaper(title, coverUrl, pathUrl);
-  //     return { error: null, data: res };
-  //   } catch (e) {
-  //     console.log(e);
-  //     return { error: e, data: null };
-  //   }
-  // }
-
   async createWallpaperNew(wallpaper: Wallpaper): Promise<ApiResult<boolean>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = (window as any).chrome.webview.hostObjects;
 
       var res = await api.CreateWallpaperNew(JSON.stringify(wallpaper));
@@ -250,7 +218,7 @@ class API {
 
   async updateWallpaper(title: string, coverUrl: string, pathUrl: string, oldWallpaper: Wallpaper): Promise<ApiResult<boolean>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = (window as any).chrome.webview.hostObjects;
 
       var res = await api.UpdateWallpaper(title, coverUrl, pathUrl, JSON.stringify(oldWallpaper));
@@ -263,7 +231,7 @@ class API {
 
   async updateWallpaperNew(wallpaper: Wallpaper, oldPath: string): Promise<ApiResult<boolean>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = (window as any).chrome.webview.hostObjects;
 
       var res = await api.UpdateWallpaperNew(JSON.stringify(wallpaper), oldPath);
@@ -276,7 +244,7 @@ class API {
 
   async deleteWallpaper(wallpaper: Wallpaper): Promise<ApiResult<boolean>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = window.chrome.webview.hostObjects;
 
       var res = await api.DeleteWallpaper(JSON.stringify(wallpaper));
@@ -289,7 +257,7 @@ class API {
 
   async explore(path: string): Promise<ApiResult<null>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = (window as any).chrome.webview.hostObjects;
 
       await api.Explore(path);
@@ -302,7 +270,7 @@ class API {
 
   async setWallpaperSetting(setting: WallpaperSetting, wallpaper: Wallpaper): Promise<ApiResult<boolean>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = (window as any).chrome.webview.hostObjects;
 
       var data = await api.SetWallpaperSetting(JSON.stringify(setting), JSON.stringify(wallpaper));
@@ -318,7 +286,7 @@ class API {
     position: number
   }>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = (window as any).chrome.webview.hostObjects;
 
       var data = await api.GetWallpaperTime(screenIndex);
@@ -331,7 +299,7 @@ class API {
 
   async setProgress(progress: number, screenIndex?: number): Promise<ApiResult<null>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = (window as any).chrome.webview.hostObjects;
 
       await api.SetProgress(progress, screenIndex);
@@ -344,7 +312,7 @@ class API {
 
   async downloadWallpaper(coverUrl: string, wallpaperUrl: string, meta: WallpaperMeta): Promise<ApiResult<boolean>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: false };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: false };
       const { api } = (window as any).chrome.webview.hostObjects;
 
       var data = await api.DownloadWallpaper(coverUrl, wallpaperUrl, JSON.stringify(meta));
@@ -357,7 +325,7 @@ class API {
 
   async cancelDownloadWallpaper(id: string): Promise<ApiResult<null>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = (window as any).chrome.webview.hostObjects;
 
       var data = await api.CancelDownloadWallpaper(id);
@@ -370,7 +338,7 @@ class API {
 
   async getDownloadItemStatus(id: string): Promise<ApiResult<DownloadItem>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = (window as any).chrome.webview.hostObjects;
       let json = await api.GetDownloadItemStatus(id);
       let data = JSON.parse(json);
@@ -382,7 +350,7 @@ class API {
   }
   async showShell(path: string): Promise<ApiResult<null>> {
     try {
-      if (!window.chrome.webview) return { error: "no webview", data: null };
+      if (!window.chrome || !window.chrome.webview) return { error: "no webview", data: null };
       const { api } = (window as any).chrome.webview.hostObjects;
       api.ShowShell(path);
       return { error: null, data: null };
@@ -406,8 +374,9 @@ class API {
     }
   }
 
+
   isRunningInClient(): boolean {
-    return typeof window !== 'undefined' && !!window.chrome.webview;
+    return typeof window !== 'undefined' && !!window.chrome?.webview;
   }
 }
